+++ HISTORY.md
@@ -1,6 +1,11 @@
 HISTORY
 ---
 
+## 0.2.1
+
+`NEW` add visual engine support
+`CHANGED` CrumbItem depends on `uxcore-crumb-item`
+
 ## 0.2.0
 `CHANGED` apply new style